♻️ refactor: 디렉토리 구성 변경 : 모듈 단위 리팩토링
