+++ src/modules/images/image.router.ts
@@ -1,7 +1,7 @@
 import express from 'express';
-import upload from '../upload';
-import { uploadSingleImage } from '../controllers/imageController';
-import { asyncHandler } from '../libs/asyncHandler';
+import upload from '../../upload';
+import { uploadSingleImage } from './image.controller';
+import { asyncHandler } from '../../libs/asyncHandler';
 
 const imageRouter = express.Router();
