+++ src/modules/notifications/notification.repository.ts
@@ -1,4 +1,4 @@
-import prisma from '../libs/prismaClient';
+import prisma from '../../libs/prismaClient';
 
 export class NotificationRepository {
   async create(userId: string, message: string) {
