+++ src/modules/products/product.repository.ts
@@ -1,6 +1,6 @@
 import { Prisma, Product } from '@prisma/client';
-import { ProductCreateDto, ProductQueryDto } from '../dto/productDto';
-import prisma from '../libs/prismaClient';
+import { ProductCreateDto, ProductQueryDto } from '../products/product.dto';
+import prisma from '../../libs/prismaClient';
 
 export class ProductRepository {
   async create(data: ProductCreateDto) {
